remove 'category' modules 'helpers' and 'macros' in lieu of domain modules
