+++ examples/calculator/features/buttons_work.feature
@@ -10,4 +10,3 @@ Feature: The calculator buttons work
   Scenario: I push some other buttons
     When the number 1 is entered
     Then the display says 1
-    Then the current tag state contains the tag "lol"
