+++ features/cuke.rs
@@ -5,7 +5,6 @@ mod step_definitions;
 mod support;
 
 use cucumber::{ WorldRunner, Server };
-use cucumber::helpers::cucumber_command;
 
 use support::env::CucumberWorld;
 use step_definitions::cucumber_steps;
@@ -20,7 +19,7 @@ fn main() {
   // NOTE: Unused stop_rx needs to be held, or it will drop and close the server
   let (handle, stop_rx) = server.start(Some("0.0.0.0:7878"));
 
-  let status = cucumber_command()
+  let status = cucumber::ruby_command()
     .spawn()
     .unwrap_or_else(|e| { panic!("failed to execute process: {}", e) })
     .wait().unwrap();
